Replace commit_files with unified file_changes stream
Single stream combining PR files + direct-push files:
- PR files: GET /pulls/{n}/files — 1 call per PR (~1K for cyberfabric)
- Direct-push files: GET /commits/{sha} — default branch non-merge only
- source_type discriminator: "pr" or "direct_push"
- pr_number nullable (null for direct pushes)

Replaces commit_files which made 1 REST call per commit (22K+ calls).
Reduces API calls from ~22K to ~1.2K for typical orgs.
